Fix structure factor comparison in tests
With the new Numpy 1.24, the way the structure factors for degenerate
modes are decomposed has changed (but they sum to the same value, so
are still correct). We were supposed to account for this in
sum_degenerate_modes, but before that was called we set the first 3
structure factors at the gamma points to zero to avoid acoustic mode
fluctuations. The problem here was in our assumption of what a gamma
point is - as we are using conversion_mat, the point 1. -1. -1. was
converted to -1.5 0.5 0.5, which is not a gamma point! So the first 3
modes were not near-zero acoustic as assumed, and in fact modes 3
and 4 were degenerate, so setting mode 3 to zero means that the summing
of degenerate modes was no longer valid, so the tests failed with
the new version of Numpy.

Instead of the gamma point check, just set to zero where the frequencies
are below 1 meV.